separate server and indexer binaries
